lab_09/requirements.txt
psycopg2-binary
redis
matplotlib
